Refactor multisign functions to new file
