+++ xrpl/transaction/__init__.py
@@ -6,20 +6,19 @@ from xrpl.asyncio.transaction import (
 from xrpl.transaction.ledger import get_transaction_from_hash
 from xrpl.transaction.main import (
     autofill,
-    multisign,
     safe_sign_and_autofill_transaction,
     safe_sign_and_submit_transaction,
     safe_sign_transaction,
-    safe_sign_transaction_with_multisign,
     submit_transaction,
 )
+from xrpl.transaction.multisign import multisign, sign_for_multisign
 from xrpl.transaction.reliable_submission import send_reliable_submission
 
 __all__ = [
     "autofill",
     "get_transaction_from_hash",
+    "sign_for_multisign",
     "safe_sign_transaction",
-    "safe_sign_transaction_with_multisign",
     "safe_sign_and_autofill_transaction",
     "safe_sign_and_submit_transaction",
     "submit_transaction",
